updated CREATE script because rowstores are better optimized

+++ tpch/quickstep/run-benchmark.sh
@@ -1,5 +1,7 @@
 # Load user-defined environment variables
 
+set -e
+
 echo "Loading settings from $1"
 if ! source $1 ; then echo "Failed to load config" ; exit 1 ; fi
 
@@ -75,10 +77,12 @@ function run_queries {
   QSEXE="$QS $QS_ARGS_BASE $QS_ARGS_NUMA_RUN $QS_ARGS_STORAGE"
   TOTALRUNS=5
   queries=( 01 02 03 04 05 06 07 08 09 10 11 12 13 14 15 16 17 18 19 20 21 22 )
-
+  echo $QSEXE
   for query in ${queries[@]} ; do
     echo "Query $query.sql"
-    rm tmp.sql &>/dev/null
+    if [ -f tmp.sql ] ; then
+      rm tmp.sql &>/dev/null
+    fi
     touch tmp.sql
     # Run each query a variable number of times.
     for i in `seq 1 $TOTALRUNS`;
@@ -94,17 +98,6 @@ function run_queries {
     then
       echo "Quickstep failed on query $query, continuing to next query."
     fi
-
-    # Run quickstep with with a timeout of 30 minutes. This is because no set of
-    # queries should run over 30 minutes.
-    # timeout 30m $QSEXE < tmp.sql
-    # if [ $? = 124 ] ;
-    # then
-    #   echo "Quickstep timed out on query $query, continuing to next query."
-    # elif [ $? != 0  ] ;
-    # then
-    #   echo "Quickstep failed on query $query, continuing to next query."
-    # fi
   done
   rm tmp.sql &>/dev/null
 }
